+++ neo4j-cypher-dsl-build/processor/pom.xml
@@ -22,25 +22,33 @@
 
 	<parent>
 		<groupId>org.neo4j</groupId>
-		<artifactId>neo4j-cypher-dsl-parent</artifactId>
+		<artifactId>neo4j-cypher-dsl-build</artifactId>
 		<version>${revision}${sha1}${changelist}</version>
 	</parent>
 
-	<artifactId>neo4j-cypher-dsl-native-tests</artifactId>
+	<artifactId>neo4j-cypher-dsl-build-proc</artifactId>
 
-	<name>Neo4j Cypher DSL (Native Tests)</name>
-	<description>Ensures Cypher DSL can be compiled and used natively.</description>
+	<name>Neo4j Cypher DSL (Build Utils: Processor)</name>
+	<description>Annotations and processors for minimum GraalVM support.</description>
 
 	<properties>
-		<java-module-name>org.neo4j.cypherdsl.graalvm</java-module-name>
-		<native-maven-plugin.version>0.9.19</native-maven-plugin.version>
+		<java-module-name>org.neo4j.cypherdsl.build.proc</java-module-name>
 	</properties>
 
 	<dependencies>
 		<dependency>
 			<groupId>org.neo4j</groupId>
-			<artifactId>neo4j-cypher-dsl-parser</artifactId>
+			<artifactId>neo4j-cypher-dsl-build-annotations</artifactId>
 			<version>${project.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-databind</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>com.google.testing.compile</groupId>
+			<artifactId>compile-testing</artifactId>
+			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.assertj</groupId>
@@ -57,41 +65,6 @@
 	<build>
 		<plugins>
 			<plugin>
-				<groupId>org.jacoco</groupId>
-				<artifactId>jacoco-maven-plugin</artifactId>
-				<configuration>
-					<skip>true</skip>
-				</configuration>
-			</plugin>
-
-			<plugin>
-				<groupId>org.graalvm.buildtools</groupId>
-				<artifactId>native-maven-plugin</artifactId>
-				<version>${native-maven-plugin.version}</version>
-				<executions>
-					<execution>
-						<phase>package</phase>
-						<goals>
-							<goal>compile-no-fork</goal>
-						</goals>
-						<configuration>
-							<requiredVersion>22.3</requiredVersion>
-							<mainClass>org.neo4j.cypherdsl.graalvm.Application</mainClass>
-							<imageName>application</imageName>
-							<buildArgs>--no-fallback -H:+ReportExceptionStackTraces</buildArgs>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
-
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-jar-plugin</artifactId>
-				<configuration>
-					<skipIfEmpty>true</skipIfEmpty>
-				</configuration>
-			</plugin>
-			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-install-plugin</artifactId>
 				<configuration>
@@ -107,21 +80,16 @@
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-failsafe-plugin</artifactId>
-				<executions>
-					<execution>
-						<goals>
-							<goal>integration-test</goal>
-							<goal>verify</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-javadoc-plugin</artifactId>
 				<configuration>
 					<skip>true</skip>
+				</configuration>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<configuration combine.self="append">
+					<compilerArgument>-proc:none</compilerArgument>
 				</configuration>
 			</plugin>
 		</plugins>
