+++ neo4j-cypher-dsl-build/processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,0 +1,1 @@
+org.neo4j.cypherdsl.build.processor.RegisterForReflectionProcessor
