+++ neo4j-cypher-dsl-build/processor/src/test/resources/test_classes/Class0.java
@@ -1,0 +1,7 @@
+package org.neo4j.cypherdsl.build.test;
+
+import org.neo4j.cypherdsl.build.annotations.RegisterForReflection;
+
+@RegisterForReflection
+public class Class0 {
+}
