+++ neo4j-cypher-dsl-build/processor/src/test/resources/test_classes/Class2.java
