+++ neo4j-cypher-dsl-build/processor/src/test/resources/test_classes/Class3.java
@@ -1,0 +1,7 @@
+package org.neo4j.cypherdsl.build.test;
+
+import org.neo4j.cypherdsl.build.annotations.RegisterForReflection;
+
+@RegisterForReflection(allDeclaredMethods = false, allDeclaredConstructors = true)
+public class Class3 {
+}
