--- neo4j-cypher-dsl-build/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,0 @@
-org.neo4j.cypherdsl.build.RegisterForReflectionProcessor
