--- neo4j-cypher-dsl-build/src/test/resources/test_classes/Class0.java
@@ -1,7 +1,0 @@
-package org.neo4j.cypherdsl.build.test;
-
-import org.neo4j.cypherdsl.build.RegisterForReflection;
-
-@RegisterForReflection
-public class Class0 {
-}
