+++ neo4j-cypher-dsl-examples/neo4j-cypher-dsl-examples-driver/pom.xml
@@ -33,7 +33,6 @@
 
 	<properties>
 		<java-module-name>org.neo4j.cypherdsl.examples.driver</java-module-name>
-		<java.version>11</java.version>
 	</properties>
 
 	<dependencyManagement>
