+++ neo4j-cypher-dsl-examples/neo4j-cypher-dsl-examples-sdn6/pom.xml
@@ -38,7 +38,6 @@
 	<properties>
 		<checkstyle.version>10.5.0</checkstyle.version>
 		<java-module-name>org.neo4j.cypherdsl.examples.sdn6</java-module-name>
-		<java.version>11</java.version>
 		<maven-checkstyle-plugin.version>3.2.0</maven-checkstyle-plugin.version>
 		<sortpom-maven-plugin.version>2.15.0</sortpom-maven-plugin.version>
 		<testcontainers.version>1.17.6</testcontainers.version>
