+++ neo4j-cypher-dsl-parser/pom.xml
@@ -34,7 +34,6 @@
 	<properties>
 		<covered-ratio-complexity>0.68</covered-ratio-complexity>
 		<java-module-name>org.neo4j.cypherdsl.parser</java-module-name>
-		<java.version>11</java.version>
 		<sonar.coverage.jacoco.xmlReportPaths>${basedir}/../${aggregate.report.dir}</sonar.coverage.jacoco.xmlReportPaths>
 	</properties>
